+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added MailChimp merge tags: SOURCE
+
 ### 1.0.6: January 30th, 2020
 * Edited MailChimp merge tags to default names: FNAME, LNAME
 * Added MailChimp merge tags: BIRTHDAY, BIRTHDATE